Move app state to own file
